+++ gossip/src/lib.rs
@@ -25,6 +25,7 @@ pub mod ping_pong;
 mod push_active_set;
 mod received_cache;
 pub mod weighted_shuffle;
+pub mod wen_restart_values;
 
 #[macro_use]
 extern crate log;
